hackaton1/requirements.txt
asgiref==3.4.1
Django==3.2.7
gunicorn==20.1.0
mysql-connector==2.2.9
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
